Feature/guest route

+++ app/Http/Controllers/CategoryController.php
@@ -15,10 +15,13 @@ class CategoryController extends Controller
      */
     public function index()
     {
-        //
-        $categories = Category::all();
-        return view('categories.index')
-            ->with('categories', $categories);
+        if (auth()->user()->type == 2) {
+            return redirect()->route('revista.index');
+        } else {
+            $categories = Category::all();
+            return view('categories.index')
+                ->with('categories', $categories);
+        }
     }
 
     /**
@@ -28,7 +31,11 @@ class CategoryController extends Controller
      */
     public function create()
     {
-        return view('categories.create');
+        if (auth()->user()->type == 2) {
+            return redirect()->route('revista.index');
+        } else {
+            return view('categories.create');
+        }
     }
 
     /**
@@ -39,24 +46,26 @@ class CategoryController extends Controller
      */
     public function store(Request $request)
     {
-        // dd($request);
-        $request->validate([
-            'img_categorias' => 'required|image|mimes:jpeg,png,jpg,gif,svg|max:2048',
-        ]);
-        $nombre = $request->input('nombre');
-        $status = $request->input('status');
-        $descripcion = $request->input('descripcion');
-        $imageName = time() . '-' . $request->input('nombre') . '.' . $request->img_categorias->extension();
-        $request->img_categorias->move(public_path('images/categorias'), $imageName);
-        DB::table('categorias')->insert([
-            'nombre' => $nombre,
-            'descripcion' => $descripcion,
-            'status' => $status,
-            'img_categorias' => $imageName,
-        ]);
-        // Category::create($request->all());
-        return back() //redirect()->route('categories.index')
-            ->with('success', 'Producto creado correctamente.');
+        if (auth()->user()->type == 2) {
+            return redirect()->route('revista.index');
+        } else {
+            $request->validate([
+                'img_categorias' => 'required|image|mimes:jpeg,png,jpg,gif,svg|max:2048',
+            ]);
+            $nombre = $request->input('nombre');
+            $status = $request->input('status');
+            $descripcion = $request->input('descripcion');
+            $imageName = time() . '-' . $request->input('nombre') . '.' . $request->img_categorias->extension();
+            $request->img_categorias->move(public_path('images/categorias'), $imageName);
+            DB::table('categorias')->insert([
+                'nombre' => $nombre,
+                'descripcion' => $descripcion,
+                'status' => $status,
+                'img_categorias' => $imageName,
+            ]);
+            return back()
+                ->with('success', 'Producto creado correctamente.');
+        }
     }
 
     /**
@@ -65,13 +74,10 @@ class CategoryController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    // public function show($id)
-    // {   
-
-    //     // $search = Category::findOrFail($id);
-    //     // return view('categories')
-    //     //     ->with('search',$search);
-    // }
+    public function show($id)
+    {
+        // 
+    }
 
     /**
      * Show the form for editing the specified resource.
@@ -81,9 +87,13 @@ class CategoryController extends Controller
      */
     public function edit($id)
     {
-        $category = Category::findOrFail($id);
-        return view('categories.edit')
-            ->with('category', $category);
+        if (auth()->user()->type == 2) {
+            return redirect()->route('revista.index');
+        } else {
+            $category = Category::findOrFail($id);
+            return view('categories.edit')
+                ->with('category', $category);
+        }
     }
 
     /**
@@ -95,36 +105,35 @@ class CategoryController extends Controller
      */
     public function update(Request $request, $id)
     {
-        if ($request->img_categorias == null) {
-            // dd($request);
-            Category::where('id', $id)
-                ->update([
-                    'nombre' => $request->input('nombre'),
-                    'descripcion' => $request->input('descripcion'),
-                    'status' => $request->input('status'),
-                ]);
-            return back()
-                ->with('success', 'Categoria editado correctamente.');
+        if (auth()->user()->type == 2) {
+            return redirect()->route('revista.index');
         } else {
-            // dd($request);
-            $request->validate([
-                // 'nombre' => 'required',
-                // 'descripcion' => 'required',
-                // 'status' => 'required',
-                'img_categorias' => 'required|image|mimes:jpeg,png,jpg,gif,svg|max:2048'
-            ]);
-            $imageName = time() . '-' . $request->nombre . '.' . $request->img_categorias->extension();
-
-            $request->img_categorias->move(public_path('images/categorias'), $imageName);
-            Category::where('id', $id)
-                ->update([
-                    'nombre' => $request->input('nombre'),
-                    'descripcion' => $request->input('descripcion'),
-                    'status' => $request->input('status'),
-                    'img_categorias' => $imageName
+            if ($request->img_categorias == null) {
+                Category::where('id', $id)
+                    ->update([
+                        'nombre' => $request->input('nombre'),
+                        'descripcion' => $request->input('descripcion'),
+                        'status' => $request->input('status'),
+                    ]);
+                return back()
+                    ->with('success', 'Categoria editado correctamente.');
+            } else {
+                $request->validate([
+                    'img_categorias' => 'required|image|mimes:jpeg,png,jpg,gif,svg|max:2048'
                 ]);
-            return back()
-                ->with('success', 'Categoria editado correctamente.');
+                $imageName = time() . '-' . $request->nombre . '.' . $request->img_categorias->extension();
+
+                $request->img_categorias->move(public_path('images/categorias'), $imageName);
+                Category::where('id', $id)
+                    ->update([
+                        'nombre' => $request->input('nombre'),
+                        'descripcion' => $request->input('descripcion'),
+                        'status' => $request->input('status'),
+                        'img_categorias' => $imageName
+                    ]);
+                return back()
+                    ->with('success', 'Categoria editado correctamente.');
+            }
         }
     }
 
@@ -136,8 +145,12 @@ class CategoryController extends Controller
      */
     public function destroy($id)
     {
-        Category::destroy($id);
-        return redirect('categories.index')
-            ->with('success', 'Categoria eliminado correctamente.');
+        if (auth()->user()->type == 2) {
+            return redirect()->route('revista.index');
+        } else {
+            Category::destroy($id);
+            return redirect('categories.index')
+                ->with('success', 'Categoria eliminado correctamente.');
+        }
     }
 }
